Move get_file_from_server to utils
