+++ clinicadl/clinicadl/preprocessing/t1_extensive/t1_extensive_utils.py
@@ -50,3 +50,43 @@ def apply_binary_mask(input_img, binary_img, output_filename):
     )
     nib.save(masked_image, masked_image_path)
     return masked_image_path
+
+
+def get_file_from_server(remote_file, cache_path=None):
+    """
+    Download file from server
+
+    Args:
+        remote_file (str): RemoteFileStructure defined in clinica.utils.inputs
+        cache_path (str): (default: ~/.cache/clinica/data)
+
+    Returns:
+        Path to downloaded file.
+
+    Note:
+        This function will be in Clinica.
+    """
+    import os
+    from pathlib import Path
+    from clinica.utils.stream import cprint
+    from clinica.utils.inputs import fetch_file
+
+    home = str(Path.home())
+    if cache_path:
+        cache_clinica = os.path.join(home, ".cache", cache_path)
+    else:
+        cache_clinica = os.path.join(home, ".cache", "clinica", "data")
+    if not (os.path.exists(cache_clinica)):
+        os.makedirs(cache_clinica)
+
+    local_file = os.path.join(cache_clinica, remote_file.filename)
+
+    if not (os.path.exists(local_file)):
+        try:
+            local_file = fetch_file(remote_file, cache_clinica)
+        except IOError as err:
+            cprint(
+                f"Unable to download {remote_file.filename} from {remote_file.url}: {err}"
+            )
+
+    return local_file
